fix print bug categorical data without formal test
A bug occurred when printing categorical data without a formal test. Two rows are provided under the colnames to allow adding the formal test. When no test is allowed, this results in two empty lines and messes up the lay-out of the table in knitr. This is solved by removing the empty row in the categorical case where no formal test is required.